# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-movingaverage",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — the moving (sliding-window) average of a list of numbers.",
        "description": "This service owns the *control flow* but delegates every arithmetic step to\nits dependencies, exactly as specified. For each start `i` in\n`0..=(values.len() - window)` it takes the window slice `values[i..i+window]`,\nasks `srvcs-sum` for its sum, then asks `srvcs-floatdivide` to divide that\nsum by `window`; the quotient is the windowed average.\n\n`window` must be `>= 1` and `<= values.len()`, otherwise the request is\n`422`. Validation of the element values themselves is propagated from\n`srvcs-sum`'s `422`.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovingAverageResponse"
                }
              }
            }
          },
          "422": {
            "description": "window out of range, or a dependency rejected an input (forwarded)"
          },
          "500": {
            "description": "a dependency returned a malformed result"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "values",
          "window"
        ],
        "properties": {
          "values": {
            "type": "object",
            "description": "The list of numbers to slide a window over."
          },
          "window": {
            "type": "integer",
            "format": "int64",
            "description": "The window size. Must be `>= 1` and `<= values.len()`."
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      },
      "MovingAverageResponse": {
        "type": "object",
        "required": [
          "values",
          "window",
          "result"
        ],
        "properties": {
          "result": {
            "type": "object",
            "description": "The list of windowed averages, as `f64`s."
          },
          "values": {
            "type": "object"
          },
          "window": {
            "type": "integer",
            "format": "int64"
          }
        }
      }
    }
  }
}